various minor changes
git-svn-id: https://svn.csail.mit.edu/drc@2850 c7283977-0100-402a-a91a-fa70b306dbfe

+++ software/perception/world_model/affordance/src/AffordanceState.cpp
@@ -238,9 +238,9 @@ void AffordanceState::initHelper(const drc::affordance_t *msg)
   memcpy(_origin_xyz, msg->origin_xyz, 3*sizeof(double));
   memcpy(_origin_rpy, msg->origin_rpy, 3*sizeof(double));
 
-  memcpy(_bounding_xyz, msg->bounding_xyz, 3*sizeof(float));
-  memcpy(_bounding_rpy, msg->bounding_rpy, 3*sizeof(float));
-  memcpy(_bounding_lwh, msg->bounding_lwh, 3*sizeof(float));
+  memcpy(_bounding_xyz, msg->bounding_xyz, 3*sizeof(double));
+  memcpy(_bounding_rpy, msg->bounding_rpy, 3*sizeof(double));
+  memcpy(_bounding_lwh, msg->bounding_lwh, 3*sizeof(double));
 
   _otdf_type 		= msg->otdf_type;
   //_points 	= msg->points;
@@ -283,9 +283,9 @@ void AffordanceState::toMsg(drc::affordance_t *msg) const
 	memcpy(msg->origin_xyz, _origin_xyz, 3*sizeof(double));
 	memcpy(msg->origin_rpy, _origin_rpy, 3*sizeof(double));
 
-	memcpy(msg->bounding_xyz, _bounding_xyz, 3*sizeof(float));
-	memcpy(msg->bounding_rpy, _bounding_rpy, 3*sizeof(float));
-	memcpy(msg->bounding_lwh, _bounding_lwh, 3*sizeof(float));
+	memcpy(msg->bounding_xyz, _bounding_xyz, 3*sizeof(double));
+	memcpy(msg->bounding_rpy, _bounding_rpy, 3*sizeof(double));
+	memcpy(msg->bounding_lwh, _bounding_lwh, 3*sizeof(double));
 
 	unordered_map<string,double>::const_iterator iter;
 	//params
